pyproject.toml
[tool.poetry]
name = "cur-convert"
version = "0.1.0"
description = ""
authors = ["Dmitriy <rubcovdmitry@maidjanfol.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
